마이페이지, 공지사항, 질문게시판, 댓글 기능 추가

+++ src/main/java/Winter_Project/Semteul_Battle/controller/Problem/ProblemController.java
@@ -1,16 +1,12 @@
 package Winter_Project.Semteul_Battle.controller.Problem;
 
 import Winter_Project.Semteul_Battle.config.jwt.JwtTokenProvider;
-import Winter_Project.Semteul_Battle.dto.Problem.AddProblemDto;
-import Winter_Project.Semteul_Battle.dto.Problem.DeleteProblemDto;
-import Winter_Project.Semteul_Battle.dto.Problem.UpdateProblemDto;
+import Winter_Project.Semteul_Battle.dto.Problem.*;
 import Winter_Project.Semteul_Battle.repository.ContestRepository;
+import Winter_Project.Semteul_Battle.repository.IORepository;
 import Winter_Project.Semteul_Battle.repository.ProblemRepository;
 import Winter_Project.Semteul_Battle.service.Contest.ContestLiveService;
-import Winter_Project.Semteul_Battle.service.Contest.ContestService;
-import Winter_Project.Semteul_Battle.service.Problem.AddProblemService;
-import Winter_Project.Semteul_Battle.service.Problem.DeleteProblemService;
-import Winter_Project.Semteul_Battle.service.Problem.UpdateProblemService;
+import Winter_Project.Semteul_Battle.service.Problem.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
@@ -23,12 +19,17 @@ public class ProblemController {
     private final AddProblemService addProblemService;
     private final JwtTokenProvider  jwtTokenProvider;
     private final DeleteProblemService deleteProblemService;
+    private final DeleteIOService deleteIOService;
     private final ProblemRepository problemRepository;
     private final ContestRepository contestRepository;
     private final ContestLiveService contestLiveService;
     private final UpdateProblemService updateProblemService;
+    private final UpdateIOService updateIOService;
+    private final IORepository ioRepository;
+    private final AddIOService addIOService;
 
-    @PostMapping("/add")
+    // 공백 문제 추가
+    @PostMapping("/addProblem")
     public boolean addProblem(@RequestBody AddProblemDto addProblemDto,
                               @RequestHeader("Authorization") String token) {
         Long contestId = addProblemDto.getContestId();
@@ -42,8 +43,24 @@ public class ProblemController {
             return false;
         }
     }
+    //공백 예시 추가
+    @PostMapping("/addIO")
+    public boolean addIO(@RequestBody AddIODto addIODto,
+                         @RequestHeader("Authorization") String token) {
+        Long contestId = addIODto.getContestId();
+        String tokenFromId = jwtTokenProvider.extractLoginIdFromToken(token); // 토큰에서 id 추출
+        Long participantStatus = contestLiveService.whoAreU(contestId, tokenFromId); // 참가자 / 출제자 구분
 
-    @DeleteMapping("/delete")
+        if (participantStatus == 0) { // 출제자가 맞는 경우
+            addIOService.IOFrame(addIODto);
+            return true;
+        } else { // 출제자가 아닌 경우
+            return false;
+        }
+    }
+
+
+    @DeleteMapping("/deleteProblem")
     public boolean deleteProblem(@RequestBody DeleteProblemDto deleteProblemDto,
                                  @RequestHeader("Authorization") String token) {
         Long contestId = deleteProblemDto.getContestId();
@@ -55,9 +72,10 @@ public class ProblemController {
             // 해당 대회와 문제가 존재하는지 확인
             boolean contestExists = contestRepository.existsById(contestId);
             boolean problemExists = problemRepository.existsById(deleteProblemDto.getProblemId());
-
+            boolean ioExists = ioRepository.existsById(deleteProblemDto.getIoId());
             // 대회와 문제가 모두 존재하는 경우에만 문제 삭제 서비스 호출
-            if (contestExists && problemExists) {
+            if (contestExists && problemExists && ioExists) {
+                deleteIOService.deleteIO(deleteProblemDto);
                 deleteProblemService.deleteProblem(deleteProblemDto);
                 return true;
             } else {
@@ -69,7 +87,7 @@ public class ProblemController {
         }
     }
 
-    @PatchMapping("/update")
+    @PatchMapping("/updateProblem")
     public boolean updateProblem(@RequestBody UpdateProblemDto updateProblemDto,
                                  @RequestHeader("Authorization") String token) {
         Long contestId = updateProblemDto.getContestId();
@@ -81,10 +99,11 @@ public class ProblemController {
             // 해당 대회와 문제가 존재하는지 확인
             boolean contestExists = contestRepository.existsById(contestId);
             boolean problemExists = problemRepository.existsById(updateProblemDto.getProblemId());
-
+            boolean ioExists = ioRepository.existsById(updateProblemDto.getIoId());
             // 대회, 문제 여부 확인 후 문제 내용 저장
-            if (contestExists && problemExists) {
-            updateProblemService.updateProblem(updateProblemDto);
+            if (contestExists && problemExists && ioExists) {
+                updateProblemService.updateProblem(updateProblemDto);
+                updateIOService.UpdateIO(updateProblemDto);
                 return true;
             } else {
                 return false;
